requirements.txt
streamlit
opencv-python-headless
Pillow
easyocr
roboflow
